Improved Panel class setup
