+++ src/Panel/Panel.php
@@ -7,14 +7,10 @@ use Kirby\Exception\Exception;
 use Kirby\Exception\InvalidArgumentException;
 use Kirby\Exception\NotFoundException;
 use Kirby\Exception\PermissionException;
-use Kirby\Filesystem\Dir;
-use Kirby\Filesystem\F;
 use Kirby\Http\Request;
 use Kirby\Http\Response;
-use Kirby\Http\Uri;
 use Kirby\Http\Url;
 use Kirby\Toolkit\A;
-use Kirby\Toolkit\Str;
 use Kirby\Toolkit\Tpl;
 use Throwable;
 
@@ -99,260 +95,6 @@ class Panel
         return $areas;
     }
 
-    /**
-     * Generates an array with all assets
-     * that need to be loaded for the panel (js, css, icons)
-     *
-     * @return array
-     */
-    public static function assets(): array
-    {
-        $kirby = kirby();
-
-        // get the assets from the Vite dev server in dev mode;
-        // dev mode = explicitly enabled in the config AND Vite is running
-        $dev   = $kirby->option('panel.dev', false);
-        $isDev = $dev !== false && is_file($kirby->roots()->panel() . '/.vite-running') === true;
-
-        if ($isDev === true) {
-            // vite on explicitly configured base URL or port 3000
-            // of the current Kirby request
-            if (is_string($dev) === true) {
-                $url = $dev;
-            } else {
-                $url = rtrim($kirby->request()->url([
-                    'port'   => 3000,
-                    'path'   => null,
-                    'params' => null,
-                    'query'  => null
-                ])->toString(), '/');
-            }
-        } else {
-            // vite is not running, use production assets
-            $url = $kirby->url('media') . '/panel/' . $kirby->versionHash();
-        }
-
-        // fetch all plugins
-        $plugins = new Plugins();
-
-        $assets = [
-            'css' => [
-                'index'   => $url . '/css/style.css',
-                'plugins' => $plugins->url('css'),
-                'custom'  => static::customCss($kirby),
-            ],
-            'icons' => [
-                'apple-touch-icon' => [
-                    'type' => 'image/png',
-                    'url'  => $url . '/apple-touch-icon.png',
-                ],
-                'shortcut icon' => [
-                    'type' => 'image/svg+xml',
-                    'url'  => $url . '/favicon.svg',
-                ],
-                'alternate icon' => [
-                    'type' => 'image/png',
-                    'url'  => $url . '/favicon.png',
-                ]
-            ],
-            'js' => [
-                'vendor'       => $url . '/js/vendor.js',
-                'pluginloader' => $url . '/js/plugins.js',
-                'plugins'      => $plugins->url('js'),
-                'custom'       => static::customJs($kirby),
-                'index'        => $url . '/js/index.js',
-            ]
-        ];
-
-        // during dev mode, add vite client and adapt
-        // path to `index.js` - vendor and stylesheet
-        // don't need to be loaded in dev mode
-        if ($isDev === true) {
-            $assets['js']['vite']   = $url . '/@vite/client';
-            $assets['js']['index']  = $url . '/src/index.js';
-            $assets['js']['vendor'] = null;
-            $assets['css']['index'] = null;
-        }
-
-        // remove missing files
-        $assets['css'] = array_filter($assets['css']);
-        $assets['js']  = array_filter($assets['js']);
-
-        return $assets;
-    }
-
-    /**
-     * Check for a custom css file from the
-     * config (panel.css)
-     *
-     * @return string|null
-     */
-    public static function customCss(): ?string
-    {
-        if ($css = kirby()->option('panel.css')) {
-            $asset = asset($css);
-
-            if ($asset->exists() === true) {
-                return $asset->url() . '?' . $asset->modified();
-            }
-        }
-
-        return null;
-    }
-
-    /**
-     * Check for a custom js file from the
-     * config (panel.js)
-     *
-     * @return string|null
-     */
-    public static function customJs(): ?string
-    {
-        if ($js = kirby()->option('panel.js')) {
-            $asset = asset($js);
-
-            if ($asset->exists() === true) {
-                return $asset->url() . '?' . $asset->modified();
-            }
-        }
-
-        return null;
-    }
-
-    /**
-     * Creates data array for Fiber and the component
-     *
-     * @param array $data
-     * @return array
-     */
-    public static function data(array $data = []): array
-    {
-        $kirby = kirby();
-
-        // multilang setup check
-        $multilang = $kirby->option('languages', false) !== false;
-
-        // shared data for all requests
-        $shared = [
-            '$language' => function () use ($kirby, $multilang) {
-                if ($multilang === true && $language = $kirby->language()) {
-                    return [
-                        'code'    => $language->code(),
-                        'default' => $language->isDefault(),
-                        'name'    => $language->name(),
-                    ];
-                }
-            },
-            '$languages' => function () use ($kirby, $multilang): array {
-                if ($multilang === true) {
-                    return $kirby->languages()->values(function ($language) {
-                        return [
-                            'code'    => $language->code(),
-                            'default' => $language->isDefault(),
-                            'name'    => $language->name(),
-                        ];
-                    });
-                }
-
-                return [];
-            },
-            '$permissions' => function () use ($kirby) {
-                if ($user = $kirby->user()) {
-                    return $user->role()->permissions()->toArray();
-                }
-            },
-            '$license' => (bool)$kirby->system()->license(),
-            '$multilang' => $multilang,
-            '$url' => Url::current(),
-            '$user' => function () use ($kirby) {
-                if ($user = $kirby->user()) {
-                    return [
-                        'email'       => $user->email(),
-                        'id'          => $user->id(),
-                        'language'    => $user->language(),
-                        'role'        => $user->role()->id(),
-                        'username'    => $user->username(),
-                    ];
-                }
-
-                return null;
-            }
-        ];
-
-        // check for explicitly requested globals to be replaced
-        $requestGlobals = $kirby->request()->header('X-Fiber-Globals') ?? get('_globals');
-
-        // split globals string into an array of fields
-        $globals = Str::split($requestGlobals, ',');
-
-        // add requested globals
-        if (empty($globals) === false) {
-            $globalsData = static::globals();
-
-            foreach ($globals as $globalsKey) {
-                if (isset($globalsData[$globalsKey]) === true) {
-                    $shared[$globalsKey] = $globalsData[$globalsKey];
-                }
-            }
-        }
-
-        // merge with shared data
-        return array_merge($shared, $data);
-    }
-
-    /**
-     * Renders the error dialog response with provided message
-     *
-     * @param string $message
-     * @param int $code
-     * @return array
-     */
-    public static function errorDialog(string $message, int $code = 404)
-    {
-        return [
-            'code'  => $code,
-            'error' => $message
-        ];
-    }
-
-    /**
-     * Renders the error view with provided message
-     *
-     * @param string $message
-     * @param int $code
-     * @return array
-     */
-    public static function errorView(string $message, int $code = 404)
-    {
-        return [
-            'code'      => $code,
-            'component' => 'k-error-view',
-            'props'     => [
-                'error'  => $message,
-                'layout' => static::hasAccess(kirby()->user()) ? 'inside' : 'outside'
-            ],
-            'title' => 'Error'
-        ];
-    }
-
-    /**
-     * Creates $fiber response array
-     *
-     * @param array $data
-     * @return array
-     */
-    public static function fiber(array $data = []): array
-    {
-        // get all data for the request
-        $data = static::data($data);
-
-        // filter data, if it's a partial request
-        $data = static::partial($data);
-
-        // resolve lazy data entries
-        return A::apply($data);
-    }
-
 
     /**
      * Check for access permissions
@@ -394,70 +136,6 @@ class Panel
         return true;
     }
 
-    /**
-     * Creates global data for the Panel.
-     * This will be injected in the full Panel
-     * view via the script tag. Global data
-     * is only requested once on the first page load.
-     * It can be loaded partially later if needed,
-     * but is otherwise not included in Fiber calls.
-     *
-     * @return array
-     */
-    public static function globals(): array
-    {
-        $kirby = kirby();
-
-        return [
-            '$config' => function () use ($kirby) {
-                return [
-                    'debug'     => $kirby->option('debug', false),
-                    'kirbytext' => $kirby->option('panel.kirbytext', true),
-                    'search'    => [
-                        'limit' => $kirby->option('panel.search.limit', 10),
-                        'type'  => $kirby->option('panel.search.type', 'pages')
-                    ],
-                    'translation' => $kirby->option('panel.language', 'en'),
-                ];
-            },
-            '$system' => function () use ($kirby) {
-                $locales = [];
-
-                foreach ($kirby->translations() as $translation) {
-                    $locales[$translation->code()] = $translation->locale();
-                }
-
-                return [
-                    'ascii'   => Str::$ascii,
-                    'csrf'    => $kirby->option('api.csrf') ?? csrf(),
-                    'isLocal' => $kirby->system()->isLocal(),
-                    'locales' => $locales,
-                    'slugs'   => Str::$language,
-                    'title'   => $kirby->site()->title()->or('Kirby Panel')->toString()
-                ];
-            },
-            '$translation' => function () use ($kirby) {
-                if ($user = $kirby->user()) {
-                    $translation = $kirby->translation($user->language());
-                } else {
-                    $translation = $kirby->translation($kirby->panelLanguage());
-                }
-
-                return [
-                    'code'      => $translation->code(),
-                    'data'      => $translation->dataWithFallback(),
-                    'direction' => $translation->direction(),
-                    'name'      => $translation->name(),
-                ];
-            },
-            '$urls' => function () use ($kirby) {
-                return [
-                    'api'  => $kirby->url('api'),
-                    'site' => $kirby->url('index')
-                ];
-            }
-        ];
-    }
 
     /**
      * Redirect to a Panel url
@@ -494,18 +172,6 @@ class Panel
     }
 
     /**
-     * Load the SVG icon sprite
-     * This will be injected in the
-     * initial HTML document for the Panel
-     *
-     * @return string
-     */
-    public static function icons(): string
-    {
-        return F::read(kirby()->root('kirby') . '/panel/dist/img/icons.svg');
-    }
-
-    /**
      * Checks for a Fiber request
      * via get parameters or headers
      *
@@ -534,86 +200,6 @@ class Panel
     {
         return Response::json($data, $code, get('_pretty'), [
             'X-Fiber' => 'true'
-        ]);
-    }
-
-    /**
-     * Links all dist files in the media folder
-     * and returns the link to the requested asset
-     *
-     * @return bool
-     * @throws \Kirby\Exception\Exception If Panel assets could not be moved to the public directory
-     */
-    public static function link(): bool
-    {
-        $kirby       = kirby();
-        $mediaRoot   = $kirby->root('media') . '/panel';
-        $panelRoot   = $kirby->root('panel') . '/dist';
-        $versionHash = $kirby->versionHash();
-        $versionRoot = $mediaRoot . '/' . $versionHash;
-
-        // check if the version already exists
-        if (is_dir($versionRoot) === true) {
-            return false;
-        }
-
-        // delete the panel folder and all previous versions
-        Dir::remove($mediaRoot);
-
-        // recreate the panel folder
-        Dir::make($mediaRoot, true);
-
-        // copy assets to the dist folder
-        if (Dir::copy($panelRoot, $versionRoot) !== true) {
-            throw new Exception('Panel assets could not be linked');
-        }
-
-        return true;
-    }
-
-    /**
-     * Filters the data array and returns only
-     * entries requested via a Fiber partial
-     * request (or the whole array if full request).
-     *
-     * @param array $data
-     * @return array
-     */
-    public static function partial(array $data): array
-    {
-        // is it a partial request?
-        $kirby       = kirby();
-        $request     = $kirby->request();
-        $requestOnly = $request->header('X-Fiber-Only') ?? get('_only');
-
-        // split include string into an array of fields
-        $only = Str::split($requestOnly, ',');
-
-        // if a full request is made, return all data
-        if (empty($only) === true) {
-            return $data;
-        }
-
-        // otherwise filter data based on
-        // dot notation, e.g. `$props.tab.columns`
-        $partials = [];
-
-        // check if globals are requested and need to be merged
-        if (Str::contains($requestOnly, '$')) {
-            $data = array_merge_recursive(static::globals(), $data);
-        }
-
-        // make sure the data is already resolved to make
-        // nested data fetching work
-        $data = A::apply($data);
-
-        // build a new array with all requested data
-        foreach ($only as $partial) {
-            $partials[$partial] = A::get($data, $partial);
-        }
-
-        return A::nest($partials, [
-            '$translation'
         ]);
     }
 
@@ -655,122 +241,10 @@ class Panel
         // handle different response types (view, dialog, ...)
         switch ($options['type'] ?? 'view') {
             case 'dialog':
-                return static::responseForDialog($result, $options);
+                return Dialog::response($result, $options);
             default:
-                return static::responseForView($result, $options);
+                return View::response($result, $options);
         }
-    }
-
-    /**
-     * Renders dialogs
-     *
-     * @param mixed $data
-     * @param array $options
-     * @return \Kirby\Http\Response
-     */
-    public static function responseForDialog($data, array $options = [])
-    {
-        // handle Kirby exceptions
-        if (is_a($data, 'Kirby\Exception\Exception') === true) {
-            $data = static::errorDialog($data->getMessage(), $data->getHttpCode());
-
-        // handle exceptions
-        } elseif (is_a($data, 'Throwable') === true) {
-            $data = static::errorDialog($data->getMessage(), 500);
-
-        // interpret true as success
-        } elseif ($data === true) {
-            $data = [
-                'code' => 200
-            ];
-
-        // only expect arrays from here on
-        } elseif (is_array($data) === false) {
-            $data = static::errorDialog('Invalid dialog response', 500);
-        }
-
-        // always inject the response code
-        $data['code']     = $data['code']    ?? 200;
-        $data['path']     = $options['path'] ?? null;
-        $data['referrer'] = static::referrer();
-
-        return static::json(['$dialog' => $data], $data['code']);
-    }
-
-    /**
-     * Renders the main panel view
-     *
-     * @param mixed $data
-     * @param array $options
-     * @return \Kirby\Http\Response
-     */
-    public static function responseForView($data, array $options = [])
-    {
-        $kirby = kirby();
-        $area  = $options['area']  ?? null;
-        $areas = $options['areas'] ?? [];
-
-        // handle Kirby exceptions
-        if (is_a($data, 'Kirby\Exception\Exception') === true) {
-            $data = static::errorView($data->getMessage(), $data->getHttpCode());
-
-        // handle regular exceptions
-        } elseif (is_a($data, 'Throwable') === true) {
-            $data = static::errorView($data->getMessage(), 500);
-
-        // only expect arrays from here on
-        } elseif (is_array($data) === false) {
-            $data = static::errorView('Invalid Panel response', 500);
-        }
-
-        // create the view based on the current area
-        $view = static::view($data, $area);
-
-        // get $fiber response array
-        $fiber = static::fiber([
-            '$view'  => $view,
-            '$areas' => array_map(function ($area) {
-                // routes and dialogs should not be included
-                // in the frontend object
-                unset($area['routes'], $area['dialogs']);
-
-                return $area;
-            }, $areas)
-        ]);
-
-        // if requested, send $fiber data as JSON
-        if (static::isFiberRequest() === true) {
-            return static::json($fiber, $view['code'] ?? 200);
-        }
-
-        // Full HTML response
-        try {
-            if (static::link() === true) {
-                usleep(1);
-                go($kirby->url('index') . '/' . $kirby->path());
-            }
-        } catch (Throwable $e) {
-            die('The Panel assets cannot be installed properly. ' . $e->getMessage());
-        }
-
-        // get the uri object for the panel url
-        $uri = new Uri($url = $kirby->url('panel'));
-
-        // inject globals
-        $globals = static::globals();
-        $fiber   = array_merge_recursive(A::apply($globals), $fiber);
-        $code    = $view['code'] ?? 200;
-
-        // load the main Panel view template
-        $body = Tpl::load($kirby->root('kirby') . '/views/panel.php', [
-            'assets'   => static::assets(),
-            'icons'    => static::icons(),
-            'nonce'    => $kirby->nonce(),
-            'fiber'    => $fiber,
-            'panelUrl' => $uri->path()->toString(true) . '/',
-        ]);
-
-        return new Response($body, 'text/html', $code);
     }
 
     /**
@@ -896,7 +370,7 @@ class Panel
         // catch all route
         $routes[] = [
             'pattern' => '(:all)',
-            'action'  => function () use ($kirby) {
+            'action'  => function () {
                 return 'The view could not be found';
             }
         ];
@@ -1019,34 +493,5 @@ class Panel
         $kirby->setCurrentTranslation($translation);
 
         return $translation;
-    }
-
-
-    /**
-     * Returns data array for view
-     *
-     * @param array $view
-     * @param array $area
-     * @return array
-     */
-    public static function view(array $view = [], ?array $area = null): array
-    {
-        $kirby = kirby();
-
-        // merge view with area defaults
-        $view = array_replace_recursive($area ?? [], $view);
-
-        $view['breadcrumb'] = $view['breadcrumb'] ?? [];
-        $view['code']       = $view['code'] ?? 200;
-        $view['path']       = Str::after($kirby->path(), '/');
-        $view['timestamp']  = (int)(microtime(true) * 1000);
-        $view['props']      = $view['props'] ?? [];
-        $view['search']     = $view['search'] ?? $kirby->option('panel.search.type', 'pages');
-
-        // make sure that routes are gone
-        unset($view['routes'], $view['dialogs']);
-
-
-        return $view;
     }
 }
