+++ tests/Panel/PanelTest.php
@@ -4,10 +4,8 @@ namespace Kirby\Panel;
 
 use Kirby\Cms\App;
 use Kirby\Filesystem\Dir;
-use Kirby\Filesystem\F;
 use Kirby\Http\Response;
 use Kirby\Toolkit\A;
-use Kirby\Toolkit\Str;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -144,394 +142,6 @@ class PanelTest extends TestCase
     }
 
     /**
-     * @covers ::assets
-     */
-    public function testAssets(): void
-    {
-        // default asset setup
-        $assets  = Panel::assets();
-        $base    = '/media/panel/' . $this->app->versionHash();
-
-        // css
-        $this->assertSame($base . '/css/style.css', $assets['css']['index']);
-        $this->assertSame('/media/plugins/index.css?0', $assets['css']['plugins']);
-
-        // icons
-        $this->assertSame($base . '/apple-touch-icon.png', $assets['icons']['apple-touch-icon']['url']);
-        $this->assertSame($base . '/favicon.svg', $assets['icons']['shortcut icon']['url']);
-        $this->assertSame($base . '/favicon.png', $assets['icons']['alternate icon']['url']);
-
-        // js
-        $this->assertSame($base . '/js/vendor.js', $assets['js']['vendor']);
-        $this->assertSame($base . '/js/plugins.js', $assets['js']['pluginloader']);
-        $this->assertSame('/media/plugins/index.js?0', $assets['js']['plugins']);
-        $this->assertSame($base . '/js/index.js', $assets['js']['index']);
-
-
-        // dev mode
-        $this->app = $this->app->clone([
-            'request' => [
-                'url' => 'http://sandbox.test'
-            ],
-            'options' => [
-                'panel' => [
-                    'dev' => true
-                ]
-            ]
-        ]);
-
-        // add vite file
-        F::write($viteFile = $this->app->roots()->panel() . '/.vite-running', '');
-
-        $assets = Panel::assets($this->app);
-        $base   = 'http://sandbox.test:3000';
-
-        // css
-        $this->assertSame(['plugins' => '/media/plugins/index.css?0'], $assets['css']);
-
-        // icons
-        $this->assertSame($base . '/apple-touch-icon.png', $assets['icons']['apple-touch-icon']['url']);
-        $this->assertSame($base . '/favicon.svg', $assets['icons']['shortcut icon']['url']);
-        $this->assertSame($base . '/favicon.png', $assets['icons']['alternate icon']['url']);
-
-        // js
-        $this->assertSame([
-            'pluginloader' => $base . '/js/plugins.js',
-            'plugins' => '/media/plugins/index.js?0',
-            'index' => $base . '/src/index.js',
-            'vite' => $base . '/@vite/client'
-        ], $assets['js']);
-
-
-        // dev mode with custom url
-        $this->app = $this->app->clone([
-            'request' => [
-                'url' => 'http://sandbox.test'
-            ],
-            'options' => [
-                'panel' => [
-                    'dev' => 'http://localhost:3000'
-                ]
-            ]
-        ]);
-
-        $assets = Panel::assets($this->app);
-        $base   = 'http://localhost:3000';
-
-        // css
-        $this->assertSame(['plugins' => '/media/plugins/index.css?0'], $assets['css']);
-
-        // icons
-        $this->assertSame($base . '/apple-touch-icon.png', $assets['icons']['apple-touch-icon']['url']);
-        $this->assertSame($base . '/favicon.svg', $assets['icons']['shortcut icon']['url']);
-        $this->assertSame($base . '/favicon.png', $assets['icons']['alternate icon']['url']);
-
-        // js
-        $this->assertSame([
-            'pluginloader' => $base . '/js/plugins.js',
-            'plugins' => '/media/plugins/index.js?0',
-            'index' => $base . '/src/index.js',
-            'vite' => $base . '/@vite/client'
-        ], $assets['js']);
-
-
-        // custom panel css and js
-        $this->app = $this->app->clone([
-            'options' => [
-                'panel' => [
-                    'css' => '/assets/panel.css',
-                    'js'  => '/assets/panel.js',
-                ]
-            ]
-        ]);
-
-        // create dummy assets
-        F::write($this->tmp . '/assets/panel.css', 'test');
-        F::write($this->tmp . '/assets/panel.js', 'test');
-
-        $assets = Panel::assets($this->app);
-
-        $this->assertTrue(Str::contains($assets['css']['custom'], 'assets/panel.css'));
-        $this->assertTrue(Str::contains($assets['js']['custom'], 'assets/panel.js'));
-
-        // clean up vite file
-        F::remove($viteFile);
-    }
-
-    /**
-     * @covers ::customCss
-     */
-    public function testCustomCss(): void
-    {
-        // invalid
-        $app = $this->app->clone([
-            'options' => [
-                'panel' => [
-                    'css' => 'nonexists.css'
-                ]
-            ]
-        ]);
-
-        $this->assertNull(Panel::customCss());
-
-        // valid
-        F::write($this->tmp . '/panel.css', '');
-
-        $app = $this->app->clone([
-            'options' => [
-                'panel' => [
-                    'css' => 'panel.css'
-                ]
-            ]
-        ]);
-
-        $this->assertTrue(Str::contains(Panel::customCss(), '/panel.css'));
-    }
-
-    /**
-     * @covers ::customJs
-     */
-    public function testCustomJs(): void
-    {
-        // invalid
-        $app = $this->app->clone([
-            'options' => [
-                'panel' => [
-                    'js' => 'nonexists.js'
-                ]
-            ]
-        ]);
-
-        $this->assertNull(Panel::customJs());
-
-        // valid
-        F::write($this->tmp . '/panel.js', '');
-
-        $app = $this->app->clone([
-            'options' => [
-                'panel' => [
-                    'js' => 'panel.js'
-                ]
-            ]
-        ]);
-
-        $this->assertTrue(Str::contains(Panel::customJs(), '/panel.js'));
-    }
-
-    /**
-     * @covers ::data
-     */
-    public function testData(): void
-    {
-        // without custom data
-        $data = Panel::data();
-
-        $this->assertInstanceOf('Closure', $data['$language']);
-        $this->assertInstanceOf('Closure', $data['$languages']);
-        $this->assertInstanceOf('Closure', $data['$permissions']);
-        $this->assertFalse($data['$license']);
-        $this->assertFalse($data['$multilang']);
-        $this->assertSame('/', $data['$url']);
-        $this->assertInstanceOf('Closure', $data['$user']);
-    }
-
-    /**
-     * @covers ::data
-     */
-    public function testDataWithIncludedGlobal(): void
-    {
-        // not included
-        $data = Panel::data();
-        $this->assertArrayNotHasKey('$translation', $data);
-
-        // via query
-        $this->app = $this->app->clone([
-            'request' => [
-                'query' => [
-                    '_globals' => '$translation'
-                ]
-            ]
-        ]);
-
-        $data = Panel::data();
-        $this->assertArrayHasKey('$translation', $data);
-
-        // via header
-        $this->app = $this->app->clone([
-            'request' => [
-                'headers' => [
-                    'X-Fiber-Globals' => '$translation'
-                ]
-            ]
-        ]);
-
-        $data = Panel::data();
-        $this->assertArrayHasKey('$translation', $data);
-    }
-
-    /**
-     * @covers ::data
-     */
-    public function testDataWithCustomProps(): void
-    {
-        $data = Panel::data([
-            '$props' => $props = [
-                'foo' => 'bar'
-            ]
-        ]);
-
-        $this->assertSame($props, $data['$props']);
-    }
-
-    /**
-     * @covers ::data
-     */
-    public function testDataWithLanguages(): void
-    {
-        $this->app = $this->app->clone([
-            'languages' => [
-                [ 'code' => 'en', 'name' => 'English', 'default' => true ],
-                [ 'code' => 'de', 'name' => 'Deutsch'],
-            ],
-            'options' => [
-                'languages' => true
-            ]
-        ]);
-
-        // without custom data
-        $data = Panel::data();
-
-        // resolve lazy data
-        $data = A::apply($data);
-
-        $this->assertTrue($data['$multilang']);
-
-        $expected = [
-            [
-                'code'    => 'en',
-                'default' => true,
-                'name'    => 'English'
-            ],
-            [
-                'code'    => 'de',
-                'default' => false,
-                'name'    => 'Deutsch'
-            ]
-        ];
-
-        $this->assertSame($expected, $data['$languages']);
-        $this->assertSame($expected[0], $data['$language']);
-    }
-
-    /**
-     * @covers ::data
-     */
-    public function testDataWithAuthenticatedUser(): void
-    {
-        // authenticate pseudo user
-        $this->app->impersonate('kirby');
-
-        // without custom data
-        $data = Panel::data();
-
-        // resolve lazy data
-        $data = A::apply($data);
-
-        // user
-        $expected = [
-            'email'    => 'kirby@getkirby.com',
-            'id'       => 'kirby',
-            'language' => 'en',
-            'role'     => 'admin',
-            'username' => 'kirby@getkirby.com'
-        ];
-
-        $this->assertSame($expected, $data['$user']);
-        $this->assertSame($this->app->user()->role()->permissions()->toArray(), $data['$permissions']);
-    }
-
-    /**
-     * @covers ::errorDialog
-     */
-    public function testErrorDialog(): void
-    {
-        // default
-        $error = Panel::errorDialog('Test');
-
-        $this->assertSame(404, $error['code']);
-        $this->assertSame('Test', $error['error']);
-
-        // custom code
-        $error = Panel::errorDialog('Test', 500);
-
-        $this->assertSame(500, $error['code']);
-        $this->assertSame('Test', $error['error']);
-    }
-
-    /**
-     * @covers ::errorView
-     */
-    public function testErrorView(): void
-    {
-        // without user
-        $error = Panel::errorView('Test');
-
-        $expected = [
-            'code' => 404,
-            'component' => 'k-error-view',
-            'props' => [
-                'error' => 'Test',
-                'layout' => 'outside'
-            ],
-            'title' => 'Error'
-        ];
-
-        $this->assertSame($expected, $error);
-
-        // with user
-        $this->app->impersonate('kirby');
-        $error = Panel::errorView('Test');
-
-        $this->assertSame('inside', $error['props']['layout']);
-
-        // user without panel access
-        $this->app->impersonate('nobody');
-        $error = Panel::errorView('Test');
-
-        $this->assertSame('outside', $error['props']['layout']);
-    }
-
-    /**
-     * @covers ::errorView
-     */
-    public function testErrorViewWithCustomCode(): void
-    {
-        $error = Panel::errorView('Test', 403);
-        $this->assertSame(403, $error['code']);
-    }
-
-    /**
-     * @covers ::fiber
-     */
-    public function testFiber(): void
-    {
-        // default
-        $fiber = Panel::fiber();
-
-        $expected = [
-            '$language' => null,
-            '$languages' => [],
-            '$permissions' => null,
-            '$license' => false,
-            '$multilang' => false,
-            '$url' => '/',
-            '$user' => null
-        ];
-
-        $this->assertSame($expected, $fiber);
-    }
-
-    /**
      * @covers ::firewall
      * @covers ::hasAccess
      */
@@ -632,75 +242,6 @@ class PanelTest extends TestCase
         Panel::firewall($app->user(), 'settings');
     }
 
-
-    /**
-     * @covers ::globals
-     */
-    public function testGlobals(): void
-    {
-        // defaults
-        $globals = Panel::globals();
-
-        $this->assertInstanceOf('Closure', $globals['$config']);
-        $this->assertInstanceOf('Closure', $globals['$system']);
-        $this->assertInstanceOf('Closure', $globals['$system']);
-        $this->assertInstanceOf('Closure', $globals['$translation']);
-        $this->assertInstanceOf('Closure', $globals['$urls']);
-
-        // defaults after apply
-        $globals     = A::apply($globals);
-        $config      = $globals['$config'];
-        $system      = $globals['$system'];
-        $translation = $globals['$translation'];
-        $urls        = $globals['$urls'];
-
-        // $config
-        $this->assertFalse($config['debug']);
-        $this->assertTrue($config['kirbytext']);
-        $this->assertSame(['limit' => 10, 'type'  => 'pages'], $config['search']);
-        $this->assertSame('en', $config['translation']);
-
-        // $system
-        $this->assertSame(Str::$ascii, $system['ascii']);
-        $this->assertSame(csrf(), $system['csrf']);
-        $this->assertFalse($system['isLocal']);
-        $this->assertArrayHasKey('de', $system['locales']);
-        $this->assertArrayHasKey('en', $system['locales']);
-        $this->assertSame('en_US', $system['locales']['en']);
-        $this->assertSame('de_DE', $system['locales']['de']);
-
-        // $translation
-        $this->assertSame('en', $translation['code']);
-        $this->assertSame($this->app->translation('en')->dataWithFallback(), $translation['data']);
-        $this->assertSame('ltr', $translation['direction']);
-        $this->assertSame('English', $translation['name']);
-
-        // $urls
-        $this->assertSame('/api', $urls['api']);
-        $this->assertSame('/', $urls['site']);
-    }
-
-    /**
-     * @covers ::globals
-     */
-    public function testGlobalsWithUser(): void
-    {
-        $this->app = $this->app->clone([
-            'users' => [
-                [
-                    'email' => 'test@getkirby.com',
-                    'language' => 'de',
-                    'role' => 'admin'
-                ]
-            ]
-        ]);
-
-        $this->app->impersonate('test@getkirby.com');
-        $globals = Panel::globals();
-        $globals = A::apply($globals);
-        $this->assertSame('de', $globals['$translation']['code']);
-    }
-
     /**
      * @covers ::go
      */
@@ -748,17 +289,6 @@ class PanelTest extends TestCase
             $this->assertSame('/foo/test', $r->getMessage());
             $this->assertSame(302, $r->getCode());
         }
-    }
-
-    /**
-     * @covers ::icons
-     */
-    public function testIcons(): void
-    {
-        $icons = Panel::icons();
-
-        $this->assertNotNull($icons);
-        $this->assertTrue(strpos($icons, '<svg', 0) !== false);
     }
 
     /**
@@ -815,174 +345,6 @@ class PanelTest extends TestCase
         $this->assertSame('application/json', $response->type());
         $this->assertSame('true', $response->header('X-Fiber'));
         $this->assertSame($data, json_decode($response->body(), true));
-    }
-
-    /**
-     * @covers ::link
-     */
-    public function testLink(): void
-    {
-        // create links
-        $link = Panel::link($this->app);
-        $this->assertTrue($link);
-
-        // try again to create links, should be return false
-        $link = Panel::link($this->app);
-        $this->assertFalse($link);
-    }
-
-    /**
-     * @covers ::partial
-     */
-    public function testPartial()
-    {
-        $data = [
-            'a' => 'A',
-            'b' => 'B'
-        ];
-
-        // default (no partial request)
-        $result = Panel::partial($data);
-
-        $this->assertSame($data, $result);
-    }
-
-    /**
-     * @covers ::partial
-     */
-    public function testPartialWithOnly(): void
-    {
-        // via get
-        $this->app = $this->app->clone([
-            'request' => [
-                'query' => [
-                    '_only' => 'a',
-                ]
-            ]
-        ]);
-
-        $data = [
-            'a' => 'A',
-            'b' => 'B'
-        ];
-
-        $result = Panel::partial($data);
-
-        $this->assertSame(['a' => 'A'], $result);
-
-        // via headers
-        $this->app = $this->app->clone([
-            'request' => [
-                'headers' => [
-                    'X-Fiber-Only' => 'a',
-                ]
-            ]
-        ]);
-
-        $data = [
-            'a' => 'A',
-            'b' => 'B'
-        ];
-
-        $result = Panel::partial($data);
-
-        $this->assertSame(['a' => 'A'], $result);
-    }
-
-    /**
-     * @covers ::partial
-     */
-    public function testPartialWithGlobal(): void
-    {
-        // simulate a simple partial request
-        $this->app = $this->app->clone([
-            'request' => [
-                'query' => [
-                    '_only' => 'a,$urls',
-                ]
-            ]
-        ]);
-
-        $data = [
-            'a' => 'A',
-            'b' => 'B'
-        ];
-
-        $result = Panel::partial($data);
-        $result = A::apply($result);
-
-        $expected = [
-            'a' => 'A',
-            '$urls' => [
-                'api' => '/api',
-                'site' => '/'
-            ]
-        ];
-
-        $this->assertSame($expected, $result);
-    }
-
-    /**
-     * @covers ::partial
-     */
-    public function testPartialWithNestedData(): void
-    {
-        // simulate a simple partial request
-        $this->app = $this->app->clone([
-            'request' => [
-                'query' => [
-                    '_only' => 'b.c',
-                ]
-            ]
-        ]);
-
-        $data = [
-            'a' => 'A',
-            'b' => [
-                'c' => 'C'
-            ]
-        ];
-
-        $result = Panel::partial($data);
-
-        $expected = [
-            'b' => [
-                'c' => 'C'
-            ]
-        ];
-
-        $this->assertSame($expected, $result);
-    }
-
-    /**
-     * @covers ::partial
-     */
-    public function testPartialWithNestedGlobal(): void
-    {
-        // simulate a simple partial request
-        $this->app = $this->app->clone([
-            'request' => [
-                'query' => [
-                    '_only' => 'a,$urls.site',
-                ]
-            ]
-        ]);
-
-        $data = [
-            'a' => 'A',
-            'b' => 'B'
-        ];
-
-        $result = Panel::partial($data);
-
-        $expected = [
-            'a' => 'A',
-            '$urls' => [
-                'site' => '/'
-            ]
-        ];
-
-        $this->assertSame($expected, $result);
     }
 
     /**
@@ -1048,212 +410,6 @@ class PanelTest extends TestCase
         $this->assertSame(500, $response->code());
         $this->assertSame('k-error-view', $json['$view']['component']);
         $this->assertSame('Test', $json['$view']['props']['error']);
-    }
-
-    /**
-     * @covers ::responseForDialog
-     */
-    public function testResponseForDialog(): void
-    {
-        $response = Panel::responseForDialog([
-            'test' => 'Test'
-        ]);
-
-        $expected = [
-            '$dialog' => [
-                'test'     => 'Test',
-                'code'     => 200,
-                'path'     => null,
-                'referrer' => '/'
-            ]
-        ];
-
-        $this->assertSame('application/json', $response->type());
-        $this->assertSame('true', $response->header('X-Fiber'));
-        $this->assertSame($expected, json_decode($response->body(), true));
-    }
-
-    /**
-     * @covers ::responseForDialog
-     */
-    public function testResponseForDialogFromTrue(): void
-    {
-        $response = Panel::responseForDialog(true);
-        $expected = [
-            '$dialog' => [
-                'code'     => 200,
-                'path'     => null,
-                'referrer' => '/'
-            ]
-        ];
-
-        $this->assertSame($expected, json_decode($response->body(), true));
-    }
-
-    /**
-     * @covers ::responseForDialog
-     */
-    public function testResponseForDialogFromInvalidData(): void
-    {
-        $response = Panel::responseForDialog(1234);
-        $expected = [
-            '$dialog' => [
-                'code'     => 500,
-                'error'    => 'Invalid dialog response',
-                'path'     => null,
-                'referrer' => '/'
-            ]
-        ];
-
-        $this->assertSame($expected, json_decode($response->body(), true));
-    }
-
-    /**
-     * @covers ::responseForDialog
-     */
-    public function testResponseForDialogFromException(): void
-    {
-        $exception = new \Exception('Test');
-        $response  = Panel::responseForDialog($exception);
-        $expected  = [
-            '$dialog' => [
-                'code'     => 500,
-                'error'    => 'Test',
-                'path'     => null,
-                'referrer' => '/'
-            ]
-        ];
-
-        $this->assertSame($expected, json_decode($response->body(), true));
-    }
-
-    /**
-     * @covers ::responseForDialog
-     */
-    public function testResponseForDialogFromKirbyException(): void
-    {
-        $exception = new \Kirby\Exception\NotFoundException('Test');
-        $response  = Panel::responseForDialog($exception);
-        $expected  = [
-            '$dialog' => [
-                'code'     => 404,
-                'error'    => 'Test',
-                'path'     => null,
-                'referrer' => '/'
-            ]
-        ];
-
-        $this->assertSame($expected, json_decode($response->body(), true));
-    }
-
-    /**
-     * @covers ::responseForView
-     */
-    public function testResponseForViewAsHTML(): void
-    {
-        // create panel dist files first to avoid redirect
-        Panel::link($this->app);
-
-        // get panel response
-        $response = Panel::responseForView([
-            'test' => 'Test'
-        ]);
-
-        $this->assertInstanceOf('\Kirby\Http\Response', $response);
-        $this->assertSame(200, $response->code());
-        $this->assertSame('text/html', $response->type());
-        $this->assertSame('UTF-8', $response->charset());
-        $this->assertNotNull($response->body());
-    }
-
-    /**
-     * @covers ::responseForView
-     */
-    public function testResponseForViewAsJson(): void
-    {
-        $this->app = $this->app->clone([
-            'request' => [
-                'query' => [
-                    '_json' => true
-                ]
-            ]
-        ]);
-
-        // get panel response
-        $response = Panel::responseForView([
-            'test' => 'Test'
-        ]);
-
-        $this->assertSame('application/json', $response->type());
-        $this->assertSame('true', $response->header('X-Fiber'));
-    }
-
-    /**
-     * @covers ::responseForView
-     */
-    public function testResponseForViewFromKirbyException()
-    {
-        // fake json request for easier assertions
-        $this->app = $this->app->clone([
-            'request' => [
-                'query' => [
-                    '_json' => true,
-                ]
-            ]
-        ]);
-
-        $exception = new \Kirby\Exception\NotFoundException('Test');
-        $response  = Panel::responseForView($exception);
-        $json      = json_decode($response->body(), true);
-
-        $this->assertSame(404, $response->code());
-        $this->assertSame('k-error-view', $json['$view']['component']);
-        $this->assertSame('Test', $json['$view']['props']['error']);
-    }
-
-    /**
-     * @covers ::responseForView
-     */
-    public function testResponseForViewFromException()
-    {
-        // fake json request for easier assertions
-        $this->app = $this->app->clone([
-            'request' => [
-                'query' => [
-                    '_json' => true,
-                ]
-            ]
-        ]);
-
-        $exception = new \Exception('Test');
-        $response  = Panel::responseForView($exception);
-        $json      = json_decode($response->body(), true);
-
-        $this->assertSame(500, $response->code());
-        $this->assertSame('k-error-view', $json['$view']['component']);
-        $this->assertSame('Test', $json['$view']['props']['error']);
-    }
-
-    /**
-     * @covers ::responseForView
-     */
-    public function testResponseForViewFromUnsupportedResult()
-    {
-        // fake json request for easier assertions
-        $this->app = $this->app->clone([
-            'request' => [
-                'query' => [
-                    '_json' => true,
-                ]
-            ]
-        ]);
-
-        $response = Panel::responseForView(1234);
-        $json     = json_decode($response->body(), true);
-
-        $this->assertSame(500, $response->code());
-        $this->assertSame('k-error-view', $json['$view']['component']);
-        $this->assertSame('Invalid Panel response', $json['$view']['props']['error']);
     }
 
     /**
@@ -1481,46 +637,5 @@ class PanelTest extends TestCase
 
         $this->assertSame('de', $translation);
         $this->assertSame('de', $this->app->translation()->code());
-    }
-
-    /**
-     * @covers ::view
-     */
-    public function testView(): void
-    {
-        // defaults
-        $result = Panel::view();
-        $expected = [
-            'breadcrumb' => [],
-            'code' => 200,
-            'path' => '',
-            'timestamp' => $result['timestamp'],
-            'props' => [],
-            'search' => 'pages'
-        ];
-
-        $this->assertSame([], $result['breadcrumb']);
-        $this->assertSame(200, $result['code']);
-        $this->assertSame('', $result['path']);
-        $this->assertArrayHaskey('timestamp', $result);
-        $this->assertSame([], $result['props']);
-        $this->assertSame('pages', $result['search']);
-
-        // with $view
-        $result = Panel::view(['search' => 'files']);
-        $this->assertSame('files', $result['search']);
-
-        // with $area
-        $result = Panel::view(['search' => 'users'], ['title' => 'Users']);
-        $expected['search'] = 'users';
-        $expected['title']  = 'Users';
-        $this->assertSame('users', $result['search']);
-        $this->assertSame('Users', $result['title']);
-
-        // make sure routes are unset
-        $result = Panel::view(['search' => 'users'], ['title' => 'Users', 'routes' => ['foo' => 'bar']]);
-        $this->assertSame('users', $result['search']);
-        $this->assertSame('Users', $result['title']);
-        $this->assertArrayNotHasKey('routes', $result);
     }
 }
